requirements.txt
# Core dependencies for Obsidian Graph Reasoner
pyyaml>=5.4.0

# Wikidata integration
SPARQLWrapper>=1.8.5
requests>=2.25.0
